Add an "output_path" key

+++ src/models/SimpleWertheim.cpp
@@ -99,7 +99,7 @@ void SimpleWertheim::set_mobility(const MultiField<double> &rho, double M0, Mult
     for(unsigned int idx = 0; idx < rho.bins(); idx++) {
         double rho_tot = rho.species_view(idx)[0];
         double my_X = _X(rho_tot);
-        double D = M0 * pow(my_X, _valence);
+        double D = M0 * std::pow(my_X, (double) _valence);
         mobility(idx, 0) = rho_tot * D; // See Dhont 1996, eq 9.32 (page 577)
     }
 }
